+++ src/tracee/tracee.h
@@ -160,7 +160,7 @@ typedef struct tracee {
 extern Tracee *get_tracee(const Tracee *tracee, pid_t pid, bool create);
 extern int inherit_config(Tracee *child, Tracee *parent, bool shared_fs);
 extern int swap_config(Tracee *tracee1, Tracee *tracee2);
-extern int parse_config(Tracee *tracee, int argc, char *argv[]);
+extern int parse_config(Tracee *tracee, size_t argc, char *argv[]);
 extern void kill_all_tracees();
 
 #endif /* TRACEE_H */
